Move memory and Java utils to Nrk.FluentCore.Environment
