+++ Natsurainko.FluentCore/Environment/MemoryUtils.cs
@@ -1,4 +1,4 @@
-﻿using Nrk.FluentCore.Classes.Datas;
+﻿using Nrk.FluentCore.Environment;
 using System;
 using System.Diagnostics;
 using System.Runtime.InteropServices;
